move dialog functions into separate file
