+++ src/handlers/fs/index.ts
@@ -13,8 +13,9 @@ import { getStudioFileFromPath } from '@/main/file'
 import { getTempScriptName } from '@/main/script'
 import { trackEvent } from '@/services/usageTracking'
 import { UsageEventName } from '@/services/usageTracking/types'
+import { showSaveDialog } from '@/utils/dialog'
 import { browserWindowFromEvent } from '@/utils/electron'
-import { readFile, showSaveDialog, writeFile } from '@/utils/fs'
+import { readFile, writeFile } from '@/utils/fs'
 import * as path from '@/utils/path'
 import { isExternalScript } from '@/utils/workspace'
